// AngularApp/angularapp.client/src/app/models/Evento.ts
import { Lote } from "./Lote";
import { PalestranteEvento } from "./PalestranteEvento";
import { RedeSocial } from "./RedeSocial";

export interface Evento {
  id: number;
  local: string;
  dataEvento?: Date;
  tema: string;
  qtdPesssoas: number;
  imagemURL: string;
  telefone: string;
  email: string;
  lote: Lote[];
  redeSocials: RedeSocial[];
  palestranteEvento: PalestranteEvento[];
}